requirements.txt
Flask
google-generativeai
python-dotenv
pdfplumber
python-docx